Rename ResourcesHolder.h to game-source-code/ResourcesHolder.h
